v.13
‣ Merged the Light Pulls and Normal Pulls profiles together.  Instead of
having two profiles, the single profile will analyze your gear and
determine whether or not it should be more cautious or not.
‣ Fixed the treasure and rare hooks so that they should now activate
properly.
‣ Given the Baleful Tokens droprate increase, the profile will now
actively use the tokens as they're collected.  This may be changed so
that the tokens are only used if you actually need the items they give.
‣ Fixed an ID for the quest: Unseen Influence.
‣ Objective Zones that have objects that can be used (Throwable Bombs,
Placeable Traps, etc) will now be used.  Bombs will be thrown on mobs,
or objective objects.  Traps will be used to root NPCs, etc.
‣ The DoWhen logics that handle treasures and rares will now determine
range by Navigator.PathDistance() rather than .Distance().  This should
prevent zig-zagging.
‣ Removed object ID 241434 (Lodged Hunting Spear) from the treasure hook
as it was causing navigational issues for users with flying mounts.